// src/types/models/spt/config/ITraderConfig.ts
import type { MinMax } from "../../../models/common/MinMax.ts";
import type { IBaseConfig } from "../../../models/spt/config/IBaseConfig.ts";
import type { LootRequest } from "../../../models/spt/services/LootRequest.ts";
export interface ITraderConfig extends IBaseConfig {
    kind: "aki-trader";
    updateTime: UpdateTime[];
    purchasesAreFoundInRaid: boolean;
    /** Should trader reset times be set based on server start time (false = bsg time - on the hour) */
    tradersResetFromServerStart: boolean;
    updateTimeDefault: number;
    traderPriceMultipler: number;
    fence: FenceConfig;
}
export interface UpdateTime {
    traderId: string;
    /** Seconds between trader resets */
    seconds: MinMax;
}
export interface FenceConfig {
    discountOptions: DiscountOptions;
    partialRefreshTimeSeconds: number;
    partialRefreshChangePercent: number;
    assortSize: number;
    weaponPresetMinMax: MinMax;
    equipmentPresetMinMax: MinMax;
    itemPriceMult: number;
    presetPriceMult: number;
    armorMaxDurabilityPercentMinMax: IItemDurabilityCurrentMax;
    weaponDurabilityPercentMinMax: IItemDurabilityCurrentMax;
    chancePlateExistsInArmorPercent: number;
    /** Key: item tpl */
    itemStackSizeOverrideMinMax: Record<string, MinMax>;
    itemTypeLimits: Record<string, number>;
    /** Prevent duplicate offers of items of specific categories by parentId */
    preventDuplicateOffersOfCategory: string[];
    regenerateAssortsOnRefresh: boolean;
    /** Max rouble price before item is not listed on flea */
    itemCategoryRoublePriceLimit: Record<string, number>;
    /** Each slotid with % to be removed prior to listing on fence */
    presetSlotsToRemoveChancePercent: Record<string, number>;
    /** Block seasonal items from appearing when season is inactive */
    blacklistSeasonalItems: boolean;
    /** Max pen value allowed to be listed on flea - affects ammo + ammo boxes */
    ammoMaxPenLimit: number;
    blacklist: string[];
    coopExtractGift: CoopExtractReward;
    btrDeliveryExpireHours: number;
}
export interface IItemDurabilityCurrentMax {
    current: MinMax;
    max: MinMax;
}
export interface CoopExtractReward extends LootRequest {
    sendGift: boolean;
    messageLocaleIds: string[];
    giftExpiryHours: number;
}
export interface DiscountOptions {
    assortSize: number;
    itemPriceMult: number;
    presetPriceMult: number;
    weaponPresetMinMax: MinMax;
    equipmentPresetMinMax: MinMax;
}